Updated Linux Install to be simpler

+++ install_linux.sh
@@ -1,23 +1,28 @@
-#!/bin/bash
-
 EXECUTEONCE=".executeonce"
-[[ -f "$EXECUTEONCE" ]] && {
-  python3 -m pip install virtualenv;
-  python3 -m virtualenv venv;
-  source venv/bin/activate;
+if [ -f "$EXECUTEONCE" ]; then
+  echo "Attempting to install virtualenv"
+  python3 -m pip install virtualenv && echo "virtualenv installed"
+
+  if [ ! -d 'venv' ]; then
+    python3 -m virtualenv venv && echo "Initialized a virtual environment"
+  fi
+  source venv/bin/activate && echo "Sourced into venv"
+
   if [ -f 'requirements.txt' ]; then
     python3 -m pip install -r requirements.txt
   else
     echo "Cannot find requirements.txt"
-    exit 1
+    return 1
   fi
-  sudo chmod u+x ./main.py || { echo "Cannot get permissions to execute" ; exit 1; }
-  rm $EXECUTEONCE;  # remove to skip this on next execute
-  echo "Setup complete.";
-} || {
+  echo "Attempting to execute 'sudo chmod u+x ./main.py' -- need password..."
+  sudo chmod u+x ./main.py || return 1
+  rm $EXECUTEONCE  # remove to skip this on next execute
+  echo "Setup complete."
+else
   # if already setup just execute
-  # ./main.py || {
-  touch "$EXECUTEONCE";  # on error, reset for next execute;
-  echo "$EXECUTEONCE" reinitialized;
-  exit 1
-}
+  ./main.py || {
+    touch "$EXECUTEONCE";  # on error, reset for next execute;
+    echo "$EXECUTEONCE" reinitialized;
+    exit 1
+  }
+fi
